% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\docType{data}
\name{base_url}
\alias{base_url}
\title{Adresse to the FT oda 'referater' FTP database}
\format{An object of class \code{character} of length 1.}
\usage{
base_url
}
\description{
Adresse to the FT oda 'referater' FTP database
}
\references{
\url{https://www.ft.dk/~/media/sites/ft/pdf/dokumenter/aabne-data/oda-browser_brugervejledning.ashx}
}
\keyword{internal}
