% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_MeetingData}
\alias{get_MeetingData}
\title{Get minutes from the Danish Parliament}
\usage{
get_MeetingData(MeetingList, return = "DTlist", verbose = FALSE)
}
\arguments{
\item{MeetingList}{list; A list of meetings you wish
to get data on buildt using \code{\link{getList_Meeting}}.}

\item{return}{character; The form to return data in. See return for more.}

\item{verbose}{Tells you what is going on.}
}
\value{
Data can be returned in one of the following formats:

\itemize{
  \item \strong{DTlist:} (default) A list containing three data.tables
  named Meeting, DagsordenPunkt and Aktivitet.
  \item \strong{DTnested:} A nested data.table, with colum
  'DagsordenPunkt' containing a data.table for each DagsordenPunkt,
  which again contains the column 'Aktivitet' containing a
  data.table for each Aktivitet
  \item \strong{XMLlist:} The raw XML data parsed as an R list
  \item \strong{XMLdata:} The raw XML data as an XMLInternalDocument
}
}
\description{
Get minutes from the Danish Parliament
}
\examples{

# Get a list of all available parliamentary sessions
FT_sessions <- getList_Session()

# Get a list of all meetings in the third session
FT_meetings <- getList_Meeting(FT_sessions[3])

# Return the parsed meeting minutes of that session as a list of data.tables
FT_data <- get_MeetingData(FT_meetings)


}
