% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getList_Meeting}
\alias{getList_Meeting}
\title{Get a list of all available parliamentary meetings}
\usage{
getList_Meeting(Samling)
}
\arguments{
\item{Samling}{character; The name of the session(s)
you wish to get a list of meetings for. Use
\code{\link{getList_Session}} to get a full list of
all available sessions.}
}
\value{
A list of character vectors with all meetings
within the given session(s).
}
\description{
Get a list of all available parliamentary meetings
}
