% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getList_Session}
\alias{getList_Session}
\title{Get a list of all available parliamentary sessions}
\usage{
getList_Session()
}
\value{
A character vector of all available parliamentary sessions
}
\description{
Get a list of all available parliamentary sessions
}
