% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_Meeting.R
\name{parse_Aktivitet}
\alias{parse_Aktivitet}
\title{Parse Aktivitet from xml data}
\usage{
parse_Aktivitet(Aktivitet)
}
\arguments{
\item{xml_dataX}{XML data containing 'Aktivitet' from 'DagsordenPunkt'}
}
\description{
Parse Aktivitet from xml data
}
\keyword{internal}
