% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_Meeting.R
\name{parse_DagsordenPunkt}
\alias{parse_DagsordenPunkt}
\title{Parse DagsordenPunkt from xml data}
\usage{
parse_DagsordenPunkt(xml_dataX)
}
\arguments{
\item{xml_dataX}{XML data containing 'DagsordenPunkt'}
}
\description{
Parse DagsordenPunkt from xml data
}
\keyword{internal}
