% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_Meeting.R
\name{parse_Meeting}
\alias{parse_Meeting}
\title{Parse FT xml data}
\usage{
parse_Meeting(xml_data, returnNested = TRUE)
}
\arguments{
\item{xml_data}{XML data}

\item{returnNested}{If TRUE (default) returns one data.table with all
data nested. If FALSE returns tree data.tables - Meeting, DagsordenPunkt
and Aktivitet.}
}
\description{
Parse FT xml data
}
\keyword{internal}
